core: Move `string::interner::WeakSet` to `utils::weak_set::WeakSet`
