+++ core/src/lib.rs
@@ -50,6 +50,7 @@ pub mod string;
 pub mod tag_utils;
 pub mod timer;
 mod types;
+pub mod utils;
 mod vminterface;
 mod xml;
